requirements.txt
streamlit==1.48.1
PyYAML==6.0.2
sentence-transformers==4.1.0
tokenizers==0.21.2
transformers @ git+https://github.com/huggingface/transformers@539c6c2fa8abc9bb97218ac0b3c3d143ba800e05
openai==1.93.0

chromadb==1.0.13
deeplake==4.3.1
mistralai==1.8.2
openai==1.93.0
numpy==2.3.0
PyYAML==6.0.2
streamlit==1.48.1
tqdm==4.67.1
Whoosh==2.7.4
intervaltree==3.1.0


sentence-transformers==4.1.0
tokenizers==0.21.2
transformers @ git+https://github.com/huggingface/transformers@539c6c2fa8abc9bb97218ac0b3c3d143ba800e05

langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8







pandas~=2.3.2
plotly~=6.3.0
pillow~=11.2.1
python-dateutil~=2.9.0.post0
transformers~=4.54.0.dev0
torch~=2.8.0
glymur~=0.14.3
requests~=2.32.4
rapidfuzz~=3.14.0